Medical & consumable quantity fixes
Maximum quantity is now a value of the base class and needed to be defined in each item which didn't previously define it.

Some items appeared to be attributed to the wrong base. Cigarettes, Cannabis Joint, and Beard Oil were moved to Consumables. EMT Bag was moved to junk.

Psi Block used Psi and Psy interchangeably, it now consistently uses Psi.

More changes to come.